% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinSeg.R
\name{BINSEG}
\alias{BINSEG}
\title{Binary Segmentation}
\usage{
BINSEG(data.vec, Kmax)
}
\arguments{
\item{data.vec}{numeric data vector of response variables[1 x n]}

\item{Kmax}{int desired number of segments}
}
\value{
Output list including the changepoint locations and the cluster assignment vector [1 x n]
}
\description{
Binary Segmentation
}
