% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DynamProg_interface}
\alias{DynamProg_interface}
\title{Optimal Segmentation}
\usage{
DynamProg_interface(data_vec, Kmax)
}
\arguments{
\item{data_vec}{numeric data vector of response variables[1 x n]}

\item{Kmax}{int desired number of segments}
}
\value{
Output list including the changepoint locations and the cluster assignment vector [1 x n]
}
\description{
Optimal Segmentation
}
