% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loss}
\alias{loss}
\title{Compute Square Loss}
\usage{
loss(sum_vec, N_vec)
}
\arguments{
\item{sum_vec}{vector of cumulatively summed input data values [1 x N]}

\item{N_vec}{vector from 1 to N [1 x N]}
}
\value{
vector loss values [1 x N]
}
\description{
Compute Square Loss
}
