% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinSeg.R
\name{BINSEG}
\alias{BINSEG}
\title{Binary Segmentation}
\usage{
BINSEG(data.dt, Kmax)
}
\arguments{
\item{data.dt}{numeric data table of response variables[A x B]}

\item{Kmax}{int desired number of segments [N]}
}
\value{
Output list including the loss values at each split[1 x N]
}
\description{
Binary Segmentation
}
